class refactoring. StandardCarport and UserDefinedCarport is now an exstention of the abstract class carport

+++ src/main/java/app/entities/Carport.java
@@ -3,12 +3,12 @@ package app.entities;
 public abstract class Carport {
    int carportID;
    String name;
-   int price;
+   float price;
    int type;
    String productionDescription;
    int specification;
 
-   Carport(int carportID, String name, int price, int type, String productionDescription, int specification) {
+   Carport(int carportID, String name, float price, int type, String productionDescription, int specification) {
        this.carportID = carportID;
        this.name = name;
        this.price = price;
